requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
